Refactored multi part upload and remove storage document

+++ controllers/document/deleteFolderFromBin.js
@@ -2,7 +2,6 @@ const createHttpError = require('http-errors');
 
 const folderSchema = require('../../models/tenant/folderSchema');
 const documentSchema = require('../../models/tenant/documentSchema');
-const storageSchema = require('../../models/tenant/storageSchema');
 
 const getTenantModel = require('../../utils/getTenantModel');
 const withTransaction = require('../../utils/withTransaction');
@@ -18,13 +17,17 @@ module.exports = async function (tenant, folderId) {
 
     const { dbName, _id: tenantId } = tenant;
 
-    const Folder = getTenantModel(dbName, 'Folder', folderSchema);
-    const Document = getTenantModel(dbName, 'Document', documentSchema);
-    const Storage = getTenantModel(dbName, 'Storage', storageSchema);
+    const Folder = getTenantModel(
+        dbName,
+        'Folder',
+        folderSchema
+    );
 
-    let totalSize = 0;
-    let deletedFiles = 0;
-    let deletedFolders = 0;
+    const Document = getTenantModel(
+        dbName,
+        'Document',
+        documentSchema
+    );
 
     const documentsToDelete = [];
     const foldersToDelete = [];
@@ -39,19 +42,19 @@ module.exports = async function (tenant, folderId) {
 
         documentsToDelete.push(...docs);
 
-        docs.forEach(doc => {
-            totalSize += doc.size || 0;
-        });
-
         const folders = await Folder.find({
             parentFolderId: parentId,
             isDeleted: true
         }).lean();
 
         for (const folder of folders) {
+
             foldersToDelete.push(folder);
+
             await collectItems(folder._id);
+
         }
+
     }
 
     await collectItems(folderId);
@@ -70,24 +73,29 @@ module.exports = async function (tenant, folderId) {
 
     foldersToDelete.push(rootFolder);
 
-    deletedFiles = documentsToDelete.length;
-    deletedFolders = foldersToDelete.length;
-
-    // Delete S3 objects first
+    // Delete S3 objects
     for (const doc of documentsToDelete) {
+
         try {
+
             await deleteObject(doc.s3Key);
+
         } catch (err) {
+
             throw new createHttpError(
                 STATUS_CODE.INTERNAL_SERVER_ERROR,
                 `Unable to delete ${doc.originalFileName} from storage.`
             );
+
         }
+
     }
 
-    const storage = await withTransaction(async (session) => {
+    // Delete MongoDB records
+    await withTransaction(async (session) => {
 
         if (documentsToDelete.length) {
+
             await Document.deleteMany(
                 {
                     _id: {
@@ -96,47 +104,54 @@ module.exports = async function (tenant, folderId) {
                 },
                 { session }
             );
+
         }
 
-        await Folder.deleteMany(
-            {
-                _id: {
-                    $in: foldersToDelete.map(folder => folder._id)
-                }
-            },
-            { session }
-        );
+        if (foldersToDelete.length) {
 
-        return await Storage.findOneAndUpdate(
-            {
-                tenantId
-            },
-            {
-                $inc: {
-                    storageUsed: -totalSize,
-                    trashedFiles: -deletedFiles,
-                    totalFolders: -deletedFolders
+            await Folder.deleteMany(
+                {
+                    _id: {
+                        $in: foldersToDelete.map(folder => folder._id)
+                    }
                 },
-                $set: {
-                    lastStorageUpdatedAt: new Date()
-                }
-            },
-            {
-                new: true,
-                session
-            }
-        ).lean();
+                { session }
+            );
+
+        }
 
     });
+
+    // Recalculate storage usage
+    const [storageStats] = await Document.aggregate([
+        {
+            $match: {
+                uploadStatus: 'uploaded',
+                isDeleted: false
+            }
+        },
+        {
+            $group: {
+                _id: null,
+                storageUsed: {
+                    $sum: '$size'
+                }
+            }
+        }
+    ]);
 
     await updateApiAnalytics({
         dbName,
         set: {
-            storageUsed: storage.storageUsed
+            storageUsed: storageStats?.storageUsed || 0
         }
     });
 
-    emitToTenant(tenantId, FOLDER_DELETED);
+    emitToTenant(
+        tenantId,
+        FOLDER_DELETED
+    );
 
     return rootFolder;
+
 };
